Split up into modules per queue type.
